+++ hyvexa-parkour-ascend/src/main/java/io/hyvexa/ascend/ui/SkillTreePage.java
@@ -38,6 +38,12 @@ public class SkillTreePage extends BaseAscendPage {
         NODE_COORDINATES.put(SkillTreeNode.RUNNER_SPEED_3, "9:1");
         NODE_COORDINATES.put(SkillTreeNode.EVOLUTION_POWER_2, "9:2");
         NODE_COORDINATES.put(SkillTreeNode.MOMENTUM_ENDURANCE, "7:2");
+        NODE_COORDINATES.put(SkillTreeNode.RUNNER_SPEED_4, "10:1");
+        NODE_COORDINATES.put(SkillTreeNode.EVOLUTION_POWER_3, "10:2");
+        NODE_COORDINATES.put(SkillTreeNode.MOMENTUM_MASTERY, "11:1");
+        NODE_COORDINATES.put(SkillTreeNode.MULTIPLIER_BOOST_2, "12:1");
+        NODE_COORDINATES.put(SkillTreeNode.ELEVATION_BOOST, "12:2");
+        NODE_COORDINATES.put(SkillTreeNode.RUNNER_SPEED_5, "13:1");
     }
 
     private static final String BUTTON_CLOSE = "Close";